Update README.md
E-Farming is the first project of my university life. Although it was a very small project but we had something extraordinary at that time. By doing this work, I got to know how to do the project, got new ideas about the project and the importance of the project, etc.